perf(webapp): componetization income view and lineChart
