+++ webapp/src/components/BarChartReport/custom-barChart.js
