+++ webapp/src/components/LineChartReport/index.js
@@ -1,8 +1,11 @@
-import React, { memo, useCallback, useState, useEffect } from 'react'
+import React, { memo, useCallback, useState } from 'react'
 import PropTypes from 'prop-types'
 import FileSaver from 'file-saver'
-import { Box } from '@mui/system'
 import { makeStyles } from '@mui/styles'
+import { useTranslation } from 'react-i18next'
+import { useCurrentPng } from 'recharts-to-png'
+import TooltipDownload from '@mui/material/Tooltip'
+import DownloadOutlined from '@mui/icons-material/DownloadOutlined'
 import {
   IconButton,
   Typography,
@@ -10,96 +13,17 @@ import {
   Switch,
   FormGroup
 } from '@mui/material'
-import TooltipDownload from '@mui/material/Tooltip'
-import DownloadOutlined from '@mui/icons-material/DownloadOutlined'
-import { useTranslation } from 'react-i18next'
-import { useCurrentPng } from 'recharts-to-png'
-import {
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  ResponsiveContainer,
-  LineChart,
-  Line,
-  Brush
-} from 'recharts'
-
-import { formatWithThousandSeparator } from '../../utils/format-with-thousand-separator'
 
 import styles from './styles'
+import CustomLineChart from './custom-lineChart'
 
 const useStyles = makeStyles(styles)
-
-const RenderChartLegend = ({ data }) => {
-  const classes = useStyles()
-  const { t } = useTranslation()
-
-  return (
-    <div className={classes.chartLegent}>
-      <a key={`key-Un${data.toLocaleLowerCase()}-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#f4d35e" />
-        {t(`un${data.toLocaleLowerCase()}`, { ns: 'generalForm' })}
-      </a>
-      <a key={`key-${data}-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#ee964b" />
-        {t(`${data.toLocaleLowerCase()}`, { ns: 'generalForm' })}
-      </a>
-      <a key={`key-total-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#19647e" />
-        {t('total', { ns: 'generalForm' })}
-      </a>
-    </div>
-  )
-}
-
-RenderChartLegend.propTypes = {
-  data: PropTypes.string
-}
-
-const CustomTooltip = ({ payload = [], label = '', coinType = '' }) => {
-  const { t } = useTranslation()
-  label = label + ''
-
-  return (
-    <div>
-      {payload &&
-        payload.map((data, i) => (
-          <div key={`${i}-tooltip`}>
-            <div>
-              {i === 0 &&
-                `${t('date', { ns: 'generalForm' })}: ${
-                  data.payload.date.split('T')[0]
-                } `}
-            </div>
-            <div>
-              {`${t('balance', {
-                ns: 'incomeRoute'
-              })}: ${formatWithThousandSeparator(
-                data.payload[data.dataKey],
-                4
-              )} ${coinType}`}
-            </div>
-          </div>
-        ))}
-    </div>
-  )
-}
-
-CustomTooltip.propTypes = {
-  payload: PropTypes.array,
-  label: PropTypes.any,
-  coinType: PropTypes.string
-}
 
 const LineChartReport = ({ data, keyTranslation, pathTranslation }) => {
   const classes = useStyles()
   const [getBarPng, { ref: lineRef }] = useCurrentPng()
   const { t } = useTranslation()
   const [selectedUSD, setSelected] = useState(false)
-  const [coinType, setCoinType] = useState('EOS')
-  const [dataKey, setDataKey] = useState('balance')
-  const width = window.innerWidth
 
   const handleChange = event => {
     setSelected(event.target.checked)
@@ -112,16 +36,6 @@ const LineChartReport = ({ data, keyTranslation, pathTranslation }) => {
       FileSaver.saveAs(png, 'line-chart.png')
     }
   }, [getBarPng])
-
-  useEffect(() => {
-    if (selectedUSD) {
-      setCoinType('USD')
-      setDataKey('usd_total')
-    } else {
-      setDataKey('balance')
-      setCoinType('EOS')
-    }
-  }, [selectedUSD])
 
   return (
     <div className={classes.root}>
@@ -147,45 +61,11 @@ const LineChartReport = ({ data, keyTranslation, pathTranslation }) => {
         </div>
       </div>
       <div className={classes.chartContainer}>
-        <ResponsiveContainer width="100%" height={300} marginTop="16px">
-          <LineChart height={300} data={data} ref={lineRef}>
-            <CartesianGrid stroke="#f0f0f0" />
-            <XAxis
-              tick={{ fontSize: 10, stroke: '#000', strokeWidth: 0.5 }}
-              dataKey="date"
-              scale="auto"
-              interval={width > 600 ? 40 : 80}
-              allowDataOverflow={false}
-            />
-            <YAxis
-              tick={{ fontSize: '10px', stroke: '#000', strokeWidth: 0.1 }}
-              scale="linear"
-            />
-            <Tooltip
-              wrapperStyle={{
-                outline: 'none',
-                borderRadius: '4px',
-                backgroundColor: '#F9F9F9',
-                fontSize: '14px',
-                padding: '8px'
-              }}
-              content={<CustomTooltip coinType={coinType} />}
-            />
-            <Line
-              type="monotone"
-              dataKey={dataKey}
-              stroke="#3866eb"
-              strokeWidth={2}
-              activeDot={{ r: 8 }}
-            />
-            <Brush
-              dataKey={'date'}
-              width={width > 600 ? width * 0.4 : 100}
-              height={20}
-              x={width > 600 ? width * 0.2 : 150}
-            />
-          </LineChart>
-        </ResponsiveContainer>
+        <CustomLineChart
+          selectedUSD={selectedUSD}
+          data={data}
+          lineRef={lineRef}
+        />
       </div>
     </div>
   )
