+++ webapp/src/routes/IncomeReport/index.js
@@ -4,7 +4,6 @@ import { useTranslation } from 'react-i18next'
 import {
   FormControl,
   FormControlLabel,
-  Typography,
   RadioGroup,
   Radio,
   Divider
@@ -15,16 +14,13 @@ import BarChartReport from '../../components/BarChartReport'
 import LineChartReport from '../../components/LineChartReport'
 import TreasuryBalance from '../../components/TreasuryBalance'
 import PieChartReport from '../../components/PieChartReport'
-import { formatWithThousandSeparator } from '../../utils'
-import TableReport from '../../components/TableReport'
 import SelectComponent from '../../components/Select'
 
 import TreasuryDisbursementsInfo from './treasury-disbursements-info'
 import styles from './styles'
+import IncomeTableReport from './income-table-report'
 
 const useStyles = makeStyles(styles)
-
-const rowsCenter = { flex: 1, align: 'center', headerAlign: 'center' }
 
 const IncomeReport = () => {
   const classes = useStyles()
@@ -51,51 +47,6 @@ const IncomeReport = () => {
 
   const tableData =
     showElectionRadio === 'allElections' ? incomeByElectionsList : delegatesList
-
-  const columns = [
-    {
-      field: tableData[0]?.name ? 'name' : 'election',
-      headerName: tableData[0]?.name
-        ? t('tableHeader1', { ns: 'incomeRoute' })
-        : t('tableElectionHeader', { ns: 'incomeRoute' }),
-      cellClassName: classes.links,
-      renderCell: param => (
-        <a
-          className={tableData[0]?.name ? '' : classes.disableLink}
-          href={`https://eosdetective.io/network/transfers?accounts=${param.value}&time_min=1661975129190&time_max=1669751129190&excludedAccounts=&excludedCategories=system`}
-        >
-          {param.value}
-        </a>
-      ),
-      ...rowsCenter
-    },
-    {
-      field: tableData[0]?.election ? 'EOS_TOTAL' : 'EOS_CLAIMED',
-      headerName: 'EOS',
-      renderCell: param => <>{formatWithThousandSeparator(param.value, 2)}</>,
-      type: 'number',
-      ...rowsCenter
-    },
-    {
-      field: tableData[0]?.election ? 'USD_TOTAL' : 'USD_CLAIMED',
-      headerName: 'USD',
-      renderCell: param => <>{formatWithThousandSeparator(param.value, 2)}</>,
-      type: 'number',
-      ...rowsCenter
-    },
-    {
-      field: 'EOS_CLAIMED_PERCENT',
-      headerName: t('tableHeader7', { ns: 'incomeRoute' }),
-      type: 'number',
-      ...rowsCenter
-    },
-    {
-      field: 'EOS_UNCLAIMED_PERCENT',
-      headerName: t('tableHeader8', { ns: 'incomeRoute' }),
-      type: 'number',
-      ...rowsCenter
-    }
-  ]
 
   return (
     <div className={classes.root}>
@@ -163,18 +114,10 @@ const IncomeReport = () => {
         typeData={'income'}
       />
       <Divider variant="middle" />
-      <div className={classes.tableContainer}>
-        <div className={classes.title}>
-          <Typography variant="span">
-            {showElectionRadio === 'allElections'
-              ? t('titleTable2', { ns: 'incomeRoute' })
-              : t('titleTable', { ns: 'incomeRoute' })}
-          </Typography>
-          <div id="id-table-container">
-            <TableReport columns={columns} dataPercent={tableData} />
-          </div>
-        </div>
-      </div>
+      <IncomeTableReport
+        tableData={tableData}
+        showElectionRadio={showElectionRadio}
+      />
     </div>
   )
 }
